# Conflicts:
#	myopia-business/business-aggregation/student-biz/src/main/java/com/wupol/myopia/business/aggregation/student/service/StudentFacade.java
#	myopia-business/business-core/system-core/src/main/java/com/wupol/myopia/business/core/system/service/TemplateDistrictService.java

+++ myopia-business/business-core/system-core/src/main/java/com/wupol/myopia/business/core/system/domain/model/Template.java
@@ -33,7 +33,7 @@ public class Template {
     /**
      * 1-档案卡 2-筛查报告
      */
-    private Byte type;
+    private Integer type;
 
     /**
      * 模板名称
